UdpListener class has been refactored to separate file
